# spec/openapi.yaml
openapi: 3.1.0
info:
  title: WikiCard Portable Collection Protocol
  version: 1.0.0
  description: |
    Reference HTTP contract for signed portable collection exports and
    WIKICARD-AUTH-CODE-1 live ownership authorization.
servers:
  - url: https://issuer.example/api
paths:
  /.well-known/wikicard-issuer.json:
    get:
      summary: Discover issuer keys and protocol capabilities
      operationId: getIssuerDocument
      responses:
        "200":
          description: Issuer discovery document
          content:
            application/json:
              schema:
                $ref: "./issuer-document.schema.json"
  /collection/export:
    get:
      summary: Export the authenticated user's signed collection
      operationId: exportCollection
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Signed collection export
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
          content:
            application/json:
              schema:
                $ref: "./collection-export.schema.json"
  /collection-transfer/authorize/preview:
    post:
      summary: Validate a browser authorization request before rendering consent
      operationId: previewCollectionAuthorization
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizationRequest"
      responses:
        "200":
          description: Validated client preview
          content:
            application/json:
              schema:
                type: object
                required: [clientId, clientHost, scope, codeExpiresIn]
                properties:
                  clientId: { type: string, format: uri }
                  clientHost: { type: string }
                  scope: { const: collection:read }
                  codeExpiresIn: { type: integer }
        "403":
          description: Client or source subject is not allowed
  /collection-transfer/authorize:
    post:
      summary: Grant one-time collection authorization after explicit user consent
      operationId: authorizeCollectionTransfer
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizationRequest"
      responses:
        "200":
          description: Redirect URL carrying a short-lived code and original state
          content:
            application/json:
              schema:
                type: object
                required: [redirectUrl, expiresIn]
                properties:
                  redirectUrl: { type: string, format: uri }
                  expiresIn: { type: integer }
  /collection-transfer/token:
    post:
      summary: Exchange a one-time authorization code using PKCE S256
      operationId: exchangeCollectionCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grant_type, code, client_id, redirect_uri, code_verifier]
              additionalProperties: false
              properties:
                grant_type: { const: authorization_code }
                code: { type: string }
                client_id: { type: string, format: uri }
                redirect_uri: { type: string, format: uri }
                code_verifier: { type: string }
      responses:
        "200":
          description: Short-lived one-time bearer token
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, expires_in, scope]
                properties:
                  access_token: { type: string }
                  token_type: { const: Bearer }
                  expires_in: { type: integer }
                  scope: { const: collection:read }
  /collection-transfer/current:
    get:
      summary: Consume a live collection access token
      operationId: getLiveCollection
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Fresh signed collection and signed import authorization
          content:
            application/json:
              schema:
                type: object
                required: [authorization, collection]
                properties:
                  authorization:
                    $ref: "./import-authorization.schema.json"
                  collection:
                    $ref: "./collection-export.schema.json"
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Replace with the host application's authenticated browser session.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    AuthorizationRequest:
      type: object
      additionalProperties: false
      required:
        - client_id
        - redirect_uri
        - response_type
        - scope
        - state
        - code_challenge
        - code_challenge_method
        - subject_id
        - destination_account
        - export_hash
      properties:
        client_id:
          type: string
          format: uri
        redirect_uri:
          type: string
          format: uri
        response_type:
          const: code
        scope:
          const: collection:read
        state:
          type: string
          minLength: 32
        code_challenge:
          type: string
        code_challenge_method:
          const: S256
        subject_id:
          type: string
        destination_account:
          type: string
        export_hash:
          type: string